+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Unreleased Changes
+
 ## Suggested release: v4.3.0
 
 #### Features supported
-Added planned support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
+This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Ethernet network
+   - Scopes
 
 # v4.2.0